+++ sbk-api/javadoc/index-all.html
@@ -353,13 +353,13 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </dd>
 <dt><a href="io/sbk/perl/impl/CompositeCSVLatencyRecorder.html" title="class in io.sbk.perl.impl"><span class="typeNameLink">CompositeCSVLatencyRecorder</span></a> - Class in <a href="io/sbk/perl/impl/package-summary.html">io.sbk.perl.impl</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="io/sbk/perl/impl/CompositeCSVLatencyRecorder.html#%3Cinit%3E(io.sbk.perl.LatencyRecordWindow,int,io.sbk.logger.Print,io.sbk.logger.Print,io.sbk.perl.ReportLatency,java.lang.String)">CompositeCSVLatencyRecorder(LatencyRecordWindow, int, Print, Print, ReportLatency, String)</a></span> - Constructor for class io.sbk.perl.impl.<a href="io/sbk/perl/impl/CompositeCSVLatencyRecorder.html" title="class in io.sbk.perl.impl">CompositeCSVLatencyRecorder</a></dt>
+<dt><span class="memberNameLink"><a href="io/sbk/perl/impl/CompositeCSVLatencyRecorder.html#%3Cinit%3E(io.sbk.perl.LatencyRecordWindow,int,io.sbk.perl.Print,io.sbk.perl.Print,io.sbk.perl.ReportLatency,java.lang.String)">CompositeCSVLatencyRecorder(LatencyRecordWindow, int, Print, Print, ReportLatency, String)</a></span> - Constructor for class io.sbk.perl.impl.<a href="io/sbk/perl/impl/CompositeCSVLatencyRecorder.html" title="class in io.sbk.perl.impl">CompositeCSVLatencyRecorder</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="io/sbk/perl/impl/CompositeHashMapLatencyRecorder.html" title="class in io.sbk.perl.impl"><span class="typeNameLink">CompositeHashMapLatencyRecorder</span></a> - Class in <a href="io/sbk/perl/impl/package-summary.html">io.sbk.perl.impl</a></dt>
 <dd>
 <div class="block">class for Performance statistics.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="io/sbk/perl/impl/CompositeHashMapLatencyRecorder.html#%3Cinit%3E(io.sbk.perl.LatencyRecordWindow,int,io.sbk.logger.Print,io.sbk.logger.Print,io.sbk.perl.ReportLatency)">CompositeHashMapLatencyRecorder(LatencyRecordWindow, int, Print, Print, ReportLatency)</a></span> - Constructor for class io.sbk.perl.impl.<a href="io/sbk/perl/impl/CompositeHashMapLatencyRecorder.html" title="class in io.sbk.perl.impl">CompositeHashMapLatencyRecorder</a></dt>
+<dt><span class="memberNameLink"><a href="io/sbk/perl/impl/CompositeHashMapLatencyRecorder.html#%3Cinit%3E(io.sbk.perl.LatencyRecordWindow,int,io.sbk.perl.Print,io.sbk.perl.Print,io.sbk.perl.ReportLatency)">CompositeHashMapLatencyRecorder(LatencyRecordWindow, int, Print, Print, ReportLatency)</a></span> - Constructor for class io.sbk.perl.impl.<a href="io/sbk/perl/impl/CompositeHashMapLatencyRecorder.html" title="class in io.sbk.perl.impl">CompositeHashMapLatencyRecorder</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="io/sbk/api/Config.html" title="class in io.sbk.api"><span class="typeNameLink">Config</span></a> - Class in <a href="io/sbk/api/package-summary.html">io.sbk.api</a></dt>
 <dd>&nbsp;</dd>
@@ -1573,7 +1573,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dd>
 <div class="block">Print the Periodic performance results.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="io/sbk/perl/LatencyWindow.html#print(long,io.sbk.logger.Print,io.sbk.perl.ReportLatencies)">print(long, Print, ReportLatencies)</a></span> - Method in class io.sbk.perl.<a href="io/sbk/perl/LatencyWindow.html" title="class in io.sbk.perl">LatencyWindow</a></dt>
+<dt><span class="memberNameLink"><a href="io/sbk/perl/LatencyWindow.html#print(long,io.sbk.perl.Print,io.sbk.perl.ReportLatencies)">print(long, Print, ReportLatencies)</a></span> - Method in class io.sbk.perl.<a href="io/sbk/perl/LatencyWindow.html" title="class in io.sbk.perl">LatencyWindow</a></dt>
 <dd>
 <div class="block">Print the window statistics.</div>
 </dd>
