+++ sbk-api/javadoc/io/sbk/api/Logger.html
@@ -187,7 +187,7 @@ extends <a href="PerformanceLogger.html" title="interface in io.sbk.api">Perform
 <code><a href="PerformanceLogger.html#addArgs(io.sbk.options.InputOptions)">addArgs</a>, <a href="PerformanceLogger.html#close(io.sbk.options.InputOptions)">close</a>, <a href="PerformanceLogger.html#getMaxLatency()">getMaxLatency</a>, <a href="PerformanceLogger.html#getMinLatency()">getMinLatency</a>, <a href="PerformanceLogger.html#getPercentiles()">getPercentiles</a>, <a href="PerformanceLogger.html#getReportingIntervalSeconds()">getReportingIntervalSeconds</a>, <a href="PerformanceLogger.html#getTimeUnit()">getTimeUnit</a>, <a href="PerformanceLogger.html#open(io.sbk.options.InputOptions,java.lang.String,io.sbk.action.Action,io.sbk.time.Time)">open</a>, <a href="PerformanceLogger.html#parseArgs(io.sbk.options.InputOptions)">parseArgs</a>, <a href="PerformanceLogger.html#printTotal(double,long,long,double,double,double,long,long,long,long,int,int,long%5B%5D)">printTotal</a>, <a href="PerformanceLogger.html#setExceptionHandler(io.sbk.exception.ExceptionHandler)">setExceptionHandler</a></code></li>
 </ul>
 <ul class="blockList">
-<li class="blockList"><a id="methods.inherited.from.class.io.sbk.logger.Print">
+<li class="blockList"><a id="methods.inherited.from.class.io.sbk.perl.Print">
 <!--   -->
 </a>
 <h3>Methods inherited from interface&nbsp;io.sbk.perl.<a href="../perl/Print.html" title="interface in io.sbk.perl">Print</a></h3>
