+++ sbk-api/javadoc/io/sbk/api/PerformanceLogger.html
@@ -252,7 +252,7 @@ extends <a href="../perl/Print.html" title="interface in io.sbk.perl">Print</a><
 </tr>
 </table>
 <ul class="blockList">
-<li class="blockList"><a id="methods.inherited.from.class.io.sbk.logger.Print">
+<li class="blockList"><a id="methods.inherited.from.class.io.sbk.perl.Print">
 <!--   -->
 </a>
 <h3>Methods inherited from interface&nbsp;io.sbk.perl.<a href="../perl/Print.html" title="interface in io.sbk.perl">Print</a></h3>
