+++ sbk-api/javadoc/io/sbk/perl/LatencyWindow.html
@@ -264,7 +264,7 @@ extends <a href="LatencyRecorder.html" title="class in io.sbk.perl">LatencyRecor
 </tr>
 <tr id="i3" class="rowColor">
 <td class="colFirst"><code>void</code></td>
-<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#print(long,io.sbk.logger.Print,io.sbk.perl.ReportLatencies)">print</a></span>&#8203;(long&nbsp;endTime,
+<th class="colSecond" scope="row"><code><span class="memberNameLink"><a href="#print(long,io.sbk.perl.Print,io.sbk.perl.ReportLatencies)">print</a></span>&#8203;(long&nbsp;endTime,
      <a href="Print.html" title="interface in io.sbk.perl">Print</a>&nbsp;logger,
      <a href="ReportLatencies.html" title="interface in io.sbk.perl">ReportLatencies</a>&nbsp;copyLatencies)</code></th>
 <td class="colLast">
@@ -408,7 +408,7 @@ extends <a href="LatencyRecorder.html" title="class in io.sbk.perl">LatencyRecor
 </dl>
 </li>
 </ul>
-<a id="print(long,io.sbk.logger.Print,io.sbk.perl.ReportLatencies)">
+<a id="print(long,io.sbk.perl.Print,io.sbk.perl.ReportLatencies)">
 <!--   -->
 </a>
 <ul class="blockList">
