+++ sbk-api/javadoc/io/sbk/perl/impl/ArrayLatencyRecorder.html
@@ -283,7 +283,7 @@ extends <a href="../LatencyRecordWindow.html" title="class in io.sbk.perl">Laten
 <!--   -->
 </a>
 <h3>Methods inherited from class&nbsp;io.sbk.perl.<a href="../LatencyWindow.html" title="class in io.sbk.perl">LatencyWindow</a></h3>
-<code><a href="../LatencyWindow.html#elapsedMilliSeconds(long)">elapsedMilliSeconds</a>, <a href="../LatencyWindow.html#getSLC(io.sbk.perl.LatencyPercentiles,int%5B%5D)">getSLC</a>, <a href="../LatencyWindow.html#print(long,io.sbk.logger.Print,io.sbk.perl.ReportLatencies)">print</a></code></li>
+<code><a href="../LatencyWindow.html#elapsedMilliSeconds(long)">elapsedMilliSeconds</a>, <a href="../LatencyWindow.html#getSLC(io.sbk.perl.LatencyPercentiles,int%5B%5D)">getSLC</a>, <a href="../LatencyWindow.html#print(long,io.sbk.perl.Print,io.sbk.perl.ReportLatencies)">print</a></code></li>
 </ul>
 <ul class="blockList">
 <li class="blockList"><a id="methods.inherited.from.class.io.sbk.perl.LatencyRecorder">
