+++ sbk-api/javadoc/io/sbk/perl/impl/CompositeHashMapLatencyRecorder.html
@@ -255,7 +255,7 @@ implements <a href="../ReportLatencies.html" title="interface in io.sbk.perl">Re
 <th class="colLast" scope="col">Description</th>
 </tr>
 <tr class="altColor">
-<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(io.sbk.perl.LatencyRecordWindow,int,io.sbk.logger.Print,io.sbk.logger.Print,io.sbk.perl.ReportLatency)">CompositeHashMapLatencyRecorder</a></span>&#8203;(<a href="../LatencyRecordWindow.html" title="class in io.sbk.perl">LatencyRecordWindow</a>&nbsp;window,
+<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(io.sbk.perl.LatencyRecordWindow,int,io.sbk.perl.Print,io.sbk.perl.Print,io.sbk.perl.ReportLatency)">CompositeHashMapLatencyRecorder</a></span>&#8203;(<a href="../LatencyRecordWindow.html" title="class in io.sbk.perl">LatencyRecordWindow</a>&nbsp;window,
                                int&nbsp;maxHashMapSizeMB,
                                <a href="../Print.html" title="interface in io.sbk.perl">Print</a>&nbsp;logger,
                                <a href="../Print.html" title="interface in io.sbk.perl">Print</a>&nbsp;loggerTotal,
@@ -338,7 +338,7 @@ implements <a href="../ReportLatencies.html" title="interface in io.sbk.perl">Re
 <!--   -->
 </a>
 <h3>Methods inherited from class&nbsp;io.sbk.perl.<a href="../LatencyWindow.html" title="class in io.sbk.perl">LatencyWindow</a></h3>
-<code><a href="../LatencyWindow.html#elapsedMilliSeconds(long)">elapsedMilliSeconds</a>, <a href="../LatencyWindow.html#getSLC(io.sbk.perl.LatencyPercentiles,int%5B%5D)">getSLC</a>, <a href="../LatencyWindow.html#print(long,io.sbk.logger.Print,io.sbk.perl.ReportLatencies)">print</a></code></li>
+<code><a href="../LatencyWindow.html#elapsedMilliSeconds(long)">elapsedMilliSeconds</a>, <a href="../LatencyWindow.html#getSLC(io.sbk.perl.LatencyPercentiles,int%5B%5D)">getSLC</a>, <a href="../LatencyWindow.html#print(long,io.sbk.perl.Print,io.sbk.perl.ReportLatencies)">print</a></code></li>
 </ul>
 <ul class="blockList">
 <li class="blockList"><a id="methods.inherited.from.class.io.sbk.perl.LatencyRecorder">
@@ -430,7 +430,7 @@ implements <a href="../ReportLatencies.html" title="interface in io.sbk.perl">Re
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a id="&lt;init&gt;(io.sbk.perl.LatencyRecordWindow,int,io.sbk.logger.Print,io.sbk.logger.Print,io.sbk.perl.ReportLatency)">
+<a id="&lt;init&gt;(io.sbk.perl.LatencyRecordWindow,int,io.sbk.perl.Print,io.sbk.perl.Print,io.sbk.perl.ReportLatency)">
 <!--   -->
 </a>
 <ul class="blockListLast">
