+++ sbk-api/src/main/java/io/sbk/logger/PerformanceLogger.java
@@ -14,6 +14,7 @@ import io.sbk.action.Action;
 import io.sbk.exception.ExceptionHandler;
 import io.sbk.options.InputOptions;
 import io.sbk.config.PerlConfig;
+import io.sbk.perl.Print;
 import io.sbk.time.Time;
 import io.sbk.time.TimeUnit;
