+++ sbk-api/src/main/java/io/sbk/logger/impl/SbkPrometheusLogger.java
@@ -16,7 +16,7 @@ import io.sbk.action.Action;
 import io.sbk.config.Config;
 import io.sbk.options.InputOptions;
 import io.sbk.logger.MetricsConfig;
-import io.sbk.logger.Print;
+import io.sbk.perl.Print;
 import io.sbk.time.Time;
 import io.sbk.system.Printer;
 import java.io.IOException;
