+++ sbk-api/src/main/java/io/sbk/perl/LatencyWindow.java
@@ -11,7 +11,6 @@
 package io.sbk.perl;
 
 import io.sbk.config.PerlConfig;
-import io.sbk.logger.Print;
 import io.sbk.time.Time;
 
 abstract public class LatencyWindow extends LatencyRecorder {
