+++ sbk-gem/javadoc/io/sbk/gem/GemLogger.html
@@ -118,7 +118,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="blockList">
 <dl>
 <dt>All Superinterfaces:</dt>
-<dd><code>io.sbk.logger.CountConnections</code>, <code><a href="GetArguments.html" title="interface in io.sbk.gem">GetArguments</a></code>, <code>io.sbk.logger.PerformanceLogger</code>, <code>io.sbk.logger.Print</code>, <code>io.sbk.logger.RamLogger</code>, <code>io.sbk.perl.ReportLatencies</code>, <code>io.sbk.logger.SetRW</code></dd>
+<dd><code>io.sbk.logger.CountConnections</code>, <code><a href="GetArguments.html" title="interface in io.sbk.gem">GetArguments</a></code>, <code>io.sbk.logger.PerformanceLogger</code>, <code>io.sbk.perl.Print</code>, <code>io.sbk.logger.RamLogger</code>, <code>io.sbk.perl.ReportLatencies</code>, <code>io.sbk.logger.SetRW</code></dd>
 </dl>
 <dl>
 <dt>All Known Implementing Classes:</dt>
@@ -162,10 +162,10 @@ extends io.sbk.logger.RamLogger, <a href="GetArguments.html" title="interface in
 <code>addArgs, close, getMaxLatency, getMinLatency, getPercentiles, getReportingIntervalSeconds, getTimeUnit, open, parseArgs, printTotal, setExceptionHandler</code></li>
 </ul>
 <ul class="blockList">
-<li class="blockList"><a id="methods.inherited.from.class.io.sbk.logger.Print">
+<li class="blockList"><a id="methods.inherited.from.class.io.sbk.perl.Print">
 <!--   -->
 </a>
-<h3>Methods inherited from interface&nbsp;io.sbk.logger.Print</h3>
+<h3>Methods inherited from interface&nbsp;io.sbk.perl.Print</h3>
 <code>print</code></li>
 </ul>
 <ul class="blockList">
