+++ sbk-gem/javadoc/io/sbk/gem/impl/SbkGemRamPrometheusLogger.html
@@ -152,7 +152,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="blockList">
 <dl>
 <dt>All Implemented Interfaces:</dt>
-<dd><code>io.sbk.logger.CountReaders</code>, <code>io.sbk.logger.CountRW</code>, <code>io.sbk.logger.CountWriters</code>, <code>io.sbk.logger.Logger</code>, <code>io.sbk.logger.PerformanceLogger</code>, <code><a href="../GemLogger.html" title="interface in io.sbk.gem">GemLogger</a></code>, <code><a href="../GetArguments.html" title="interface in io.sbk.gem">GetArguments</a></code>, <code>io.sbk.logger.Print</code>, <code>io.sbk.perl.ReportLatencies</code>, <code>io.sbk.perl.ReportLatency</code>, <code>io.sbk.logger.CountConnections</code>, <code>io.sbk.logger.RamLogger</code>, <code>io.sbk.logger.SetRW</code></dd>
+<dd><code>io.sbk.logger.CountReaders</code>, <code>io.sbk.logger.CountRW</code>, <code>io.sbk.logger.CountWriters</code>, <code>io.sbk.logger.Logger</code>, <code>io.sbk.logger.PerformanceLogger</code>, <code><a href="../GemLogger.html" title="interface in io.sbk.gem">GemLogger</a></code>, <code><a href="../GetArguments.html" title="interface in io.sbk.gem">GetArguments</a></code>, <code>io.sbk.perl.Print</code>, <code>io.sbk.perl.ReportLatencies</code>, <code>io.sbk.perl.ReportLatency</code>, <code>io.sbk.logger.CountConnections</code>, <code>io.sbk.logger.RamLogger</code>, <code>io.sbk.logger.SetRW</code></dd>
 </dl>
 <hr>
 <pre>public class <span class="typeNameLabel">SbkGemRamPrometheusLogger</span>
@@ -298,10 +298,10 @@ implements <a href="../GemLogger.html" title="interface in io.sbk.gem">GemLogger
 <code>addArgs, close, getMaxLatency, getMinLatency, getPercentiles, getReportingIntervalSeconds, getTimeUnit, open, parseArgs, printTotal, setExceptionHandler</code></li>
 </ul>
 <ul class="blockList">
-<li class="blockList"><a id="methods.inherited.from.class.io.sbk.logger.Print">
+<li class="blockList"><a id="methods.inherited.from.class.io.sbk.perl.Print">
 <!--   -->
 </a>
-<h3>Methods inherited from interface&nbsp;io.sbk.logger.Print</h3>
+<h3>Methods inherited from interface&nbsp;io.sbk.perl.Print</h3>
 <code>print</code></li>
 </ul>
 <ul class="blockList">
