+++ sbk-gem/javadoc/io/sbk/gem/package-tree.html
@@ -155,7 +155,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </li>
 </ul>
 </li>
-<li class="circle">io.sbk.logger.Print
+<li class="circle">io.sbk.perl.Print
 <ul>
 <li class="circle">io.sbk.logger.PerformanceLogger
 <ul>
