+++ sbk-gem/javadoc/overview-tree.html
@@ -192,7 +192,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 </li>
 </ul>
 </li>
-<li class="circle">io.sbk.logger.Print
+<li class="circle">io.sbk.perl.Print
 <ul>
 <li class="circle">io.sbk.logger.PerformanceLogger
 <ul>
