+++ sbk-ram/javadoc/index-all.html
@@ -253,7 +253,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <dl>
 <dt><a href="io/sbk/ram/impl/RamBenchmark.html" title="class in io.sbk.ram.impl"><span class="typeNameLink">RamBenchmark</span></a> - Class in <a href="io/sbk/ram/impl/package-summary.html">io.sbk.ram.impl</a></dt>
 <dd>&nbsp;</dd>
-<dt><span class="memberNameLink"><a href="io/sbk/ram/impl/RamBenchmark.html#%3Cinit%3E(int,int,io.sbk.perl.LatencyRecordWindow,io.sbk.time.Time,int,io.sbk.perl.ReportLatencies,io.sbk.logger.SetRW,io.sbk.logger.Print)">RamBenchmark(int, int, LatencyRecordWindow, Time, int, ReportLatencies, SetRW, Print)</a></span> - Constructor for class io.sbk.ram.impl.<a href="io/sbk/ram/impl/RamBenchmark.html" title="class in io.sbk.ram.impl">RamBenchmark</a></dt>
+<dt><span class="memberNameLink"><a href="io/sbk/ram/impl/RamBenchmark.html#%3Cinit%3E(int,int,io.sbk.perl.LatencyRecordWindow,io.sbk.time.Time,int,io.sbk.perl.ReportLatencies,io.sbk.logger.SetRW,io.sbk.perl.Print)">RamBenchmark(int, int, LatencyRecordWindow, Time, int, ReportLatencies, SetRW, Print)</a></span> - Constructor for class io.sbk.ram.impl.<a href="io/sbk/ram/impl/RamBenchmark.html" title="class in io.sbk.ram.impl">RamBenchmark</a></dt>
 <dd>&nbsp;</dd>
 <dt><a href="io/sbk/ram/RamConfig.html" title="class in io.sbk.ram"><span class="typeNameLink">RamConfig</span></a> - Class in <a href="io/sbk/ram/package-summary.html">io.sbk.ram</a></dt>
 <dd>&nbsp;</dd>
