+++ sbk-ram/javadoc/io/sbk/ram/RamLogger.html
@@ -118,7 +118,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="blockList">
 <dl>
 <dt>All Superinterfaces:</dt>
-<dd><code><a href="CountConnections.html" title="interface in io.sbk.ram">CountConnections</a></code>, <code>io.sbk.logger.PerformanceLogger</code>, <code>io.sbk.logger.Print</code>, <code>io.sbk.perl.ReportLatencies</code>, <code><a href="SetRW.html" title="interface in io.sbk.ram">SetRW</a></code></dd>
+<dd><code><a href="CountConnections.html" title="interface in io.sbk.ram">CountConnections</a></code>, <code>io.sbk.logger.PerformanceLogger</code>, <code>io.sbk.perl.Print</code>, <code>io.sbk.perl.ReportLatencies</code>, <code><a href="SetRW.html" title="interface in io.sbk.ram">SetRW</a></code></dd>
 </dl>
 <dl>
 <dt>All Known Implementing Classes:</dt>
@@ -155,10 +155,10 @@ extends io.sbk.logger.PerformanceLogger, io.sbk.perl.ReportLatencies, <a href="S
 <code>addArgs, close, getMaxLatency, getMinLatency, getPercentiles, getReportingIntervalSeconds, getTimeUnit, open, parseArgs, printTotal, setExceptionHandler</code></li>
 </ul>
 <ul class="blockList">
-<li class="blockList"><a id="methods.inherited.from.class.io.sbk.logger.Print">
+<li class="blockList"><a id="methods.inherited.from.class.io.sbk.perl.Print">
 <!--   -->
 </a>
-<h3>Methods inherited from interface&nbsp;io.sbk.logger.Print</h3>
+<h3>Methods inherited from interface&nbsp;io.sbk.perl.Print</h3>
 <code>print</code></li>
 </ul>
 <ul class="blockList">
