+++ sbk-ram/javadoc/io/sbk/ram/impl/RamBenchmark.html
@@ -158,14 +158,14 @@ implements io.sbk.api.Benchmark, <a href="../RamRegistry.html" title="interface 
 <th class="colLast" scope="col">Description</th>
 </tr>
 <tr class="altColor">
-<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(int,int,io.sbk.perl.LatencyRecordWindow,io.sbk.time.Time,int,io.sbk.perl.ReportLatencies,io.sbk.logger.SetRW,io.sbk.logger.Print)">RamBenchmark</a></span>&#8203;(int&nbsp;maxQueue,
+<th class="colConstructorName" scope="row"><code><span class="memberNameLink"><a href="#%3Cinit%3E(int,int,io.sbk.perl.LatencyRecordWindow,io.sbk.time.Time,int,io.sbk.perl.ReportLatencies,io.sbk.logger.SetRW,io.sbk.perl.Print)">RamBenchmark</a></span>&#8203;(int&nbsp;maxQueue,
             int&nbsp;idleMS,
             io.sbk.perl.LatencyRecordWindow&nbsp;window,
             io.sbk.time.Time&nbsp;time,
             int&nbsp;reportingIntervalMS,
             io.sbk.perl.ReportLatencies&nbsp;reportLatencies,
             <a href="../SetRW.html" title="interface in io.sbk.ram">SetRW</a>&nbsp;setRW,
-            io.sbk.logger.Print&nbsp;logger)</code></th>
+            io.sbk.perl.Print&nbsp;logger)</code></th>
 <td class="colLast">&nbsp;</td>
 </tr>
 </table>
@@ -230,7 +230,7 @@ implements io.sbk.api.Benchmark, <a href="../RamRegistry.html" title="interface 
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a id="&lt;init&gt;(int,int,io.sbk.perl.LatencyRecordWindow,io.sbk.time.Time,int,io.sbk.perl.ReportLatencies,io.sbk.logger.SetRW,io.sbk.logger.Print)">
+<a id="&lt;init&gt;(int,int,io.sbk.perl.LatencyRecordWindow,io.sbk.time.Time,int,io.sbk.perl.ReportLatencies,io.sbk.logger.SetRW,io.sbk.perl.Print)">
 <!--   -->
 </a>
 <ul class="blockListLast">
@@ -243,7 +243,7 @@ implements io.sbk.api.Benchmark, <a href="../RamRegistry.html" title="interface 
                     int&nbsp;reportingIntervalMS,
                     io.sbk.perl.ReportLatencies&nbsp;reportLatencies,
                     <a href="../SetRW.html" title="interface in io.sbk.ram">SetRW</a>&nbsp;setRW,
-                    io.sbk.logger.Print&nbsp;logger)</pre>
+                    io.sbk.perl.Print&nbsp;logger)</pre>
 </li>
 </ul>
 </li>
