+++ sbk-ram/javadoc/io/sbk/ram/impl/SbkRamPrometheusLogger.html
@@ -147,7 +147,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="blockList">
 <dl>
 <dt>All Implemented Interfaces:</dt>
-<dd><code>io.sbk.logger.CountReaders</code>, <code>io.sbk.logger.CountRW</code>, <code>io.sbk.logger.CountWriters</code>, <code>io.sbk.logger.Logger</code>, <code>io.sbk.logger.PerformanceLogger</code>, <code>io.sbk.logger.Print</code>, <code>io.sbk.perl.ReportLatencies</code>, <code>io.sbk.perl.ReportLatency</code>, <code><a href="../CountConnections.html" title="interface in io.sbk.ram">CountConnections</a></code>, <code><a href="../RamLogger.html" title="interface in io.sbk.ram">RamLogger</a></code>, <code><a href="../SetRW.html" title="interface in io.sbk.ram">SetRW</a></code></dd>
+<dd><code>io.sbk.logger.CountReaders</code>, <code>io.sbk.logger.CountRW</code>, <code>io.sbk.logger.CountWriters</code>, <code>io.sbk.logger.Logger</code>, <code>io.sbk.logger.PerformanceLogger</code>, <code>io.sbk.perl.Print</code>, <code>io.sbk.perl.ReportLatencies</code>, <code>io.sbk.perl.ReportLatency</code>, <code><a href="../CountConnections.html" title="interface in io.sbk.ram">CountConnections</a></code>, <code><a href="../RamLogger.html" title="interface in io.sbk.ram">RamLogger</a></code>, <code><a href="../SetRW.html" title="interface in io.sbk.ram">SetRW</a></code></dd>
 </dl>
 <hr>
 <pre>public class <span class="typeNameLabel">SbkRamPrometheusLogger</span>
@@ -512,7 +512,7 @@ implements <a href="../SetRW.html" title="interface in io.sbk.ram">SetRW</a>, <a
                   long[]&nbsp;percentileValues)</pre>
 <dl>
 <dt><span class="overrideSpecifyLabel">Specified by:</span></dt>
-<dd><code>print</code>&nbsp;in interface&nbsp;<code>io.sbk.logger.Print</code></dd>
+<dd><code>print</code>&nbsp;in interface&nbsp;<code>io.sbk.perl.Print</code></dd>
 <dt><span class="overrideSpecifyLabel">Overrides:</span></dt>
 <dd><code>print</code>&nbsp;in class&nbsp;<code>io.sbk.logger.impl.SbkPrometheusLogger</code></dd>
 </dl>
