+++ sbk-ram/javadoc/io/sbk/ram/impl/package-tree.html
@@ -105,7 +105,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">java.lang.Object
 <ul>
-<li class="circle">io.sbk.logger.impl.MetricsLogger (implements io.sbk.logger.Print)
+<li class="circle">io.sbk.logger.impl.MetricsLogger (implements io.sbk.perl.Print)
 <ul>
 <li class="circle">io.sbk.logger.impl.RWMetricsLogger (implements io.sbk.logger.CountRW)
 <ul>
