+++ sbk-ram/javadoc/io/sbk/ram/package-tree.html
@@ -123,7 +123,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="circle">io.sbk.ram.<a href="RamParameterOptions.html" title="interface in io.sbk.ram"><span class="typeNameLink">RamParameterOptions</span></a> (also extends io.sbk.ram.<a href="RamParameters.html" title="interface in io.sbk.ram">RamParameters</a>)</li>
 </ul>
 </li>
-<li class="circle">io.sbk.logger.Print
+<li class="circle">io.sbk.perl.Print
 <ul>
 <li class="circle">io.sbk.logger.PerformanceLogger
 <ul>
