+++ sbk-ram/javadoc/overview-tree.html
@@ -107,7 +107,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <ul>
 <li class="circle">java.lang.Object
 <ul>
-<li class="circle">io.sbk.logger.impl.MetricsLogger (implements io.sbk.logger.Print)
+<li class="circle">io.sbk.logger.impl.MetricsLogger (implements io.sbk.perl.Print)
 <ul>
 <li class="circle">io.sbk.logger.impl.RWMetricsLogger (implements io.sbk.logger.CountRW)
 <ul>
@@ -165,7 +165,7 @@ $('.navPadding').css('padding-top', $('.fixedNav').css("height"));
 <li class="circle">io.sbk.ram.<a href="io/sbk/ram/RamParameterOptions.html" title="interface in io.sbk.ram"><span class="typeNameLink">RamParameterOptions</span></a> (also extends io.sbk.ram.<a href="io/sbk/ram/RamParameters.html" title="interface in io.sbk.ram">RamParameters</a>)</li>
 </ul>
 </li>
-<li class="circle">io.sbk.logger.Print
+<li class="circle">io.sbk.perl.Print
 <ul>
 <li class="circle">io.sbk.logger.PerformanceLogger
 <ul>
